break commands into their own dir
